requirements.txt
asgiref==3.7.2
Django==4.2.4
django-markdown-deux==1.0.6
importlib-metadata==6.8.0
Markdown==3.4.4
markdown2==2.4.10
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
zipp==3.16.2
